+++ components/Link.tsx
@@ -5,10 +5,10 @@ import setup from '../src/setup/Link.ts';
  * @todo [!] Add hover animation.
  */
 export default function (props: Partial<iLink>) {
-  const { c, children, ...p } = setup(props);
+  const { c, fref, children, ...p } = setup(props);
 
   return (
-    <a {...p} class={c.link}>
+    <a ref={fref} class={c.link} {...p}>
       {children}
     </a>
   );