+++ components/Separator.tsx
@@ -5,7 +5,7 @@ import setup from '../src/setup/Separator.ts';
  * @todo [!] Add a gradient property that creates a custom gradient separator.
  */
 export default function (props: Partial<iSeparator>) {
-  const { c, ...p } = setup(props);
+  const { c, fref, ...p } = setup(props);
 
-  return <hr {...p} class={c.separator}></hr>;
+  return <hr ref={fref} class={c.separator} {...p} />;
 }