+++ components/Text.tsx
@@ -1,21 +1,12 @@
-import { iText } from "../src/types/props.ts";
-import setup from "../src/setup/Text.ts";
+import { iText } from '../src/types/props.ts';
+import setup from '../src/setup/Text.ts';
 
 export default function (props: Partial<iText>) {
-  const {
-    c,
-    type,
-    compact,
-    single,
-    indent,
-    noMargins,
-    children,
-    inheritColor,
-    ...p
-  } = setup(props);
+  const { c, fref, type, compact, single, indent, noMargins, children, inheritColor, ...p } =
+    setup(props);
 
   return (
-    <span {...p} class={c.span}>
+    <span ref={fref} class={c.span} {...p}>
       {children}
     </span>
   );